login redirect

+++ www/admin/courses/delete.php
@@ -3,6 +3,13 @@ require '../../../vendor/autoload.php';
 include '../../../autoload.php';
 
 use controllers\CourseController;
+use core\helpers\HTTP;
+
+if (!isset($_SESSION['admin'])) {
+    HTTP::redirect("/login");
+    exit();
+}
+
 
 $courseController = new CourseController();
 $courseController->deleteCourse($_GET['id']);
