+++ www/admin/departments/delete.php
@@ -3,6 +3,13 @@ require '../../../vendor/autoload.php';
 include '../../../autoload.php';
 
 use controllers\DeptController;
+use core\helpers\HTTP;
+
+if (!isset($_SESSION['admin'])) {
+    HTTP::redirect("/login");
+    exit();
+}
+
 
 $deptController = new DeptController();
 $deptController->deleteDept($_GET['id']);
